complete reorganization of modules
