+++ docs/_backup/index.rst
@@ -3,20 +3,17 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to rhapsody's documentation!
+Welcome to Rhapsody's documentation!
 ====================================
+
 .. toctree::
    :maxdepth: 4
- 
-.. automodule:: rhapsody
- 
-.. autoclass:: rhapsody
-    :members:
- 
+   :caption: Contents
+
+   rhapsody
+
 Indices and tables
 ==================
- 
+
 * :ref:`genindex`
 * :ref:`modindex`
-* :ref:`search`
-
