+++ docs/_backup/modules.rst
@@ -1,0 +1,7 @@
+rhapsody
+========
+
+.. toctree::
+   :maxdepth: 4
+
+   rhapsody
