+++ docs/rhapsody.features.rst
@@ -1,0 +1,59 @@
+rhapsody.features package
+=========================
+
+.. automodule:: rhapsody.features
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+rhapsody.features.BLOSUM module
+-------------------------------
+
+.. automodule:: rhapsody.features.BLOSUM
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+rhapsody.features.EVmutation module
+-----------------------------------
+
+.. automodule:: rhapsody.features.EVmutation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+rhapsody.features.PDB module
+----------------------------
+
+.. automodule:: rhapsody.features.PDB
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+rhapsody.features.Pfam module
+-----------------------------
+
+.. automodule:: rhapsody.features.Pfam
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+rhapsody.features.PolyPhen2 module
+----------------------------------
+
+.. automodule:: rhapsody.features.PolyPhen2
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+rhapsody.features.Uniprot module
+--------------------------------
+
+.. automodule:: rhapsody.features.Uniprot
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
