+++ docs/rhapsody.predict.rst
@@ -1,0 +1,35 @@
+rhapsody.predict package
+========================
+
+.. automodule:: rhapsody.predict
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+rhapsody.predict.core module
+----------------------------
+
+.. automodule:: rhapsody.predict.core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+rhapsody.predict.figures module
+-------------------------------
+
+.. automodule:: rhapsody.predict.figures
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+rhapsody.predict.main module
+----------------------------
+
+.. automodule:: rhapsody.predict.main
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
