+++ docs/rhapsody.rst
@@ -2,106 +2,16 @@ rhapsody package
 ================
 
 .. automodule:: rhapsody
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 Subpackages
 -----------
 
 .. toctree::
 
-    rhapsody.test_module
-
-Submodules
-----------
-
-rhapsody.EVmutation module
---------------------------
-
-.. automodule:: rhapsody.EVmutation
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-rhapsody.PDB module
--------------------
-
-.. automodule:: rhapsody.PDB
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-rhapsody.PolyPhen2 module
--------------------------
-
-.. automodule:: rhapsody.PolyPhen2
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-rhapsody.RFtraining module
---------------------------
-
-.. automodule:: rhapsody.RFtraining
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-rhapsody.Uniprot module
------------------------
-
-.. automodule:: rhapsody.Uniprot
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-rhapsody.calcFeatures module
-----------------------------
-
-.. automodule:: rhapsody.calcFeatures
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-rhapsody.figures module
------------------------
-
-.. automodule:: rhapsody.figures
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-rhapsody.interfaces module
---------------------------
-
-.. automodule:: rhapsody.interfaces
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-rhapsody.main module
---------------------
-
-.. automodule:: rhapsody.main
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-rhapsody.rhapsody module
-------------------------
-
-.. automodule:: rhapsody.rhapsody
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-rhapsody.settings module
-------------------------
-
-.. automodule:: rhapsody.settings
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
+   rhapsody.features
+   rhapsody.predict
+   rhapsody.train
+   rhapsody.utils
