--- docs/rhapsody.test_module.rst
@@ -1,20 +1,0 @@
-rhapsody.test\_module package
-=============================
-
-.. automodule:: rhapsody.test_module
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-Submodules
-----------
-
-rhapsody.test\_module.test\_submodule module
---------------------------------------------
-
-.. automodule:: rhapsody.test_module.test_submodule
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
