+++ docs/rhapsody.train.rst
@@ -1,0 +1,27 @@
+rhapsody.train package
+======================
+
+.. automodule:: rhapsody.train
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+rhapsody.train.RFtraining module
+--------------------------------
+
+.. automodule:: rhapsody.train.RFtraining
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+rhapsody.train.figures module
+-----------------------------
+
+.. automodule:: rhapsody.train.figures
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
