+++ docs/rhapsody.utils.rst
@@ -1,0 +1,19 @@
+rhapsody.utils package
+======================
+
+.. automodule:: rhapsody.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+rhapsody.utils.settings module
+------------------------------
+
+.. automodule:: rhapsody.utils.settings
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
