--- rhapsody/new_index
@@ -1,50 +1,0 @@
-utils
-	settings
-		initialSetup
-		getDefaultClassifiers
-		delSettings
-		getSettings
-	miscellaneous
-		seqScanning
-		printSAVlist
-
-train
-	RFtraining
-		calcMetrics
-		calcPathogenicityProbs
-		RandomForestCV
-		trainRFclassifier
-	figures
-		print_pred_distrib_figure
-		print_path_prob_figure
-		print_ROC_figure
-		print_feat_imp_figure
-
-predict
-	core
-		Rhapsody class
-		buildFeatMatrix(featset, all_features)
-		calcPredictions(feat_matrix, clsf, SAV_coords=None)
-	main
-		rhapsody
-	figures
-		print_sat_mutagen_figure
-
-features
-	Uniprot
-		UniprotMapping class
-		mapSAVs2PDB(SAV_coords, custom_PDB=None)
-	PDB
-		PDBfeatures class
-		calcPDBfeatures(mapped_SAVs, sel_feats=None, custom_PDB=None)
-	PolyPhen2
-		queryPolyPhen2
-		parsePolyPhen2output
-		getSAVcoords
-		calcPolyPhen2features
-	EVmutation
-		recoverEVmutFeatures
-	Pfam
-		calcPfamFeatures
-	BLOSUM
-		calcBLOSUMfeatures
