--- rhapsody/old_index
@@ -1,51 +1,0 @@
-core
-	settings
-		initialSetup
-		getDefaultClassifiers
-		delSettings
-		getSettings
-	rhapsody
-		class
-	miscellaneous
-		seqScanning
-		printSAVlist
-		mapSAVs2PDB
-		calcPredictions
-
-train
-	RFtraining
-		calcMetrics
-		calcPathogenicityProbs
-		RandomForestCV
-		trainRFclassifier
-	figures
-		print_pred_distrib_figure
-		print_path_prob_figure
-		print_ROC_figure
-		print_feat_imp_figure
-
-predict
-	main
-		rhapsody
-	figures
-		print_sat_mutagen_figure
-
-features
-	PDB
-		class
-	calcFeatures
-		calcPolyPhen2features
-		calcPDBfeatures
-		calcPfamFeatures
-		calcBLOSUMfeatures
-		buildFeatMatrix
-
-databases
-	Uniprot
-		class
-	PolyPhen2
-		queryPolyPhen2
-		parsePolyPhen2output
-		getSAVcoords
-	EVmutation
-		recoverEVmutFeatures
