--- rhapsody/test_module/__init__.py
@@ -1,5 +1,0 @@
-__all__ = []
-
-from . import test_submodule
-from .test_submodule import *
-__all__.extend(test_submodule.__all__)
