--- rhapsody/test_module/test_submodule.py
@@ -1,7 +1,0 @@
-__all__ = ['test_function_1', 'test_function_2']
-
-def test_function_1(arg):
-    return None
-    
-def test_function_2(arg):
-    return None
