+++ rhapsody/train/__init__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+"""This subpackage contains modules for training Rhapsody classifiers and
+assess their accuracy.
+"""
+
+__all__ = []
+
+from . import RFtraining
+from .RFtraining import *
+__all__.extend(RFtraining.__all__)
+__all__.append('RFtraining')
+
+from . import figures
+from .figures import *
+__all__.extend(figures.__all__)
+__all__.append('figures')
